requirements.txt
aiofiles
apscheduler
dateparser
fastapi
google-api-core
google-genai
httpx
itsdangerous
jinja2
neo4j
numpy
pydantic
pydantic-settings>=2.0.0
python-dotenv
python-multipart
pytz
pyyaml
qdrant-client>=1.7.0,<2.0.0
redis[asyncio]>=5.0.0,<6.0.0
uvicorn
